abstract diagnosis generation
